Move settings from sidebar to main page
