+++ components/global/Navbar/SiteLogo.tsx
@@ -8,7 +8,7 @@ const calsans = localFont({
 export function SiteLogo() {
   return (
     <div className='flex items-center gap-2'>
-      <span className={cn(calsans.className, 'text-2xl ml-4')}>z planner</span>
+      <span className={cn(calsans.className, 'text-xl ml-4')}>z planner</span>
       <span className='text-slate-600 tracking-widest'>BETA</span>
     </div>
   )
