+++ components/planner/SettingsView/ManageBoardsDialog/AddNewBoardDialogContent.tsx
@@ -1,5 +1,8 @@
 import { DialogContent, DialogDescription, DialogHeader, DialogTitle } from '@/components/ui/dialog'
+import { Quicksand } from 'next/font/google'
 import { AddNewBoardForm } from './AddNewBoardForm'
+
+const quicksand = Quicksand({ subsets: ['latin'], weight: ['300', '400', '500', '600', '700'] })
 
 type AddNewCategoryDialogContentProps = {
   closeDialog: () => void
@@ -7,7 +10,7 @@ type AddNewCategoryDialogContentProps = {
 
 export const AddNewBoardDialogContent = ({ closeDialog }: AddNewCategoryDialogContentProps) => {
   return (
-    <DialogContent>
+    <DialogContent className={quicksand.className}>
       <DialogHeader>
         <DialogTitle>Add New Board</DialogTitle>
         <DialogDescription className='pt-4'>
