+++ components/planner/SettingsView/ManageBoardsDialog/DeleteBoardConfirmDialog.tsx
