+++ components/planner/Sidebar/ManageCategoriesCard/CategoryColorPicker.tsx
