+++ components/planner/Sidebar/ManageCategoriesCard/DeleteCategoryConfirmDialog.tsx
