+++ components/planner/Sidebar/ManageCategoriesCard/ModifyCategoryDialogContent.tsx
