+++ hooks/Planner/types.tsx
@@ -80,7 +80,7 @@ export type BoardsType = {
 
 export type PlannerType = {
   hasLoaded: boolean
-  currentView: 'board' | 'archive'
+  currentView: 'board' | 'archive' | 'settings'
   backendErrorOccurred: boolean
   selectedBoard: string
   isSubTaskBeingDragged: boolean
