+++ package.json
@@ -35,6 +35,7 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cmdk": "^0.2.0",
+    "cypress": "^13.13.0",
     "date-fns": "^2.30.0",
     "dotenv": "^16.3.1",
     "framer-motion": "^11.2.11",
